Add files via upload
Magazine article about the measurement and calculations surrounding light.  The article discusses the three common measurement methodologies: total flux, illuminance, and luminance. It is not rigorous but meant to be generic enough for use with a simple calculator.  It allows for "back of the envelope" estimations of the lighting required for various situations.